# Optimized `TreeNode.ancestors`