This PR adds parameter `include_self` to `TreeNode.ancestors` to optionally include the initial node in the path (default: False). This logic is consistent with the newly added `path`, which has `include_ends`. With this parameter, one can easily generate a lineage from any taxon to the root.

@christapo I guess this addition can also make `path` simpler. What do you think?

This PR also improved the performance of `ancestors`. In my benchmarks:

Before: 40.3 ms ± 565 μs
After: 21.8 ms ± 194 μs
After (with `include_self=True`): 22.1 ms ± 278 μs

Please complete the following checklist:

* [x] I have read the [contribution guidelines](https://scikit.bio/contribute.html).

* [x] I have documented all public-facing changes in the [changelog](https://github.com/scikit-bio/scikit-bio/blob/main/CHANGELOG.md).

* [ ] **This pull request includes code, documentation, or other content derived from external source(s).** If this is the case, ensure the external source's license is compatible with scikit-bio's [license](https://github.com/scikit-bio/scikit-bio/blob/main/LICENSE.txt). Include the license in the `licenses` directory and add a comment in the code giving proper attribution. Ensure any other requirements set forth by the license and/or author are satisfied.

  - **It is your responsibility to disclose** code, documentation, or other content derived from external source(s). If you have questions about whether something can be included in the project or how to give proper attribution, include those questions in your pull request and a reviewer will assist you.

* [x] This pull request does not include code, documentation, or other content derived from external source(s).

Note: [This document](https://scikit.bio/devdoc/review.html) may also be helpful to see some of the things code reviewers will be verifying when reviewing your pull request.