@christapo 2024-09-25
I think this option is a good idea! I think it will simplify the`path` function as well as other potential functions that will utilize `TreeNode.ancestors`.